LayoutTests/fast/dom/nested-timer-invisible-element-throttling.html

<!DOCTYPE html>
<body>
<script src="../../resources/js-test-pre.js"></script>
<div id="testElement" style="display: none">Test</div>
<script>
description("Tests that a nested timer changing the style of a non-visible element gets throttled.");
jsTestIsAsync = true;

var iterationCount = 0;
var timeoutId;
var testElement = document.getElementById("testElement");
var wasThrottled = false;

function checkTimerThrottleState()
{
  if (iterationCount == 5) {
    // Do not use shouldBeTrue() here because it would cause a DOM tree mutation and
    // unthrottle the DOM Timer.
    wasThrottled = internals.isTimerThrottled(timeoutId);
  } else if (iterationCount == 6) {
    debug("5th iteration, timer should have been throttled.");
    shouldBeTrue("wasThrottled");
    debug("6th iteration, timer should still be throttled.");
    shouldBeTrue("internals.isTimerThrottled(timeoutId)");
  } else if (iterationCount >= 6) {
    // Timer should be unthrottled due to the DOM tree mutation caused by shouldBeTrue() in
    // the previous iteration.
    debug("" + iterationCount + "th iteration, timer should no longer be throttled.");
    shouldBeFalse("internals.isTimerThrottled(timeoutId)");
    clearTimeout(timeoutId);
    finishJSTest();
  }
}

function timerCallback()
{
  ++iterationCount;
  // Interact with the style of the element.
  testElement.style["left"] = "" + iterationCount + "px";

  // 5 iterations should be sufficient to throttle the timer.
  if (iterationCount >= 5) {
    setTimeout(checkTimerThrottleState, 0);
    if (iterationCount == 7) {
      // The call to debug() mutates the DOM tree.
      debug("Timer mutated the DOM tree.");
    }
  }

  timeoutId = setTimeout(timerCallback, 0);
}

timeoutId = setTimeout(timerCallback, 0);
debug("The timer should initially not be throttled.");
shouldBeFalse("internals.isTimerThrottled(timeoutId)");
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
